This commit implements saving of meta data of wiki links that are included in a SVGs and the rendering of wiki links through cache files.
The references of wiki links are stored both in the meta file of the SVG and the indexer.
This is a preparation step for implement the rewrite of wiki links inside the SVG when pages are moved with the move plugin.

Also a render function is implemented that renders wiki links inside the SVG and stores the rendered SVG in the cache.

This is a preliminary version - the meta data is updated on each fetch of the status of the SVG (very inefficent).

This commit is part of lejmr/dokuwiki-plugin-drawio#61